feat: add reusable workflow to build openedx images  (#1)
Add reusable workflow to build Open edX images based on the Picasso V2 pipeline definition. The main goal of this implementation is to prove the feasibility of adopting Github Actions to replace Jenkins for our current build process. Please visit our announcement post for more details on the team's stand on the potential migration.