refactor: drop TVM for cleaner implementation in isolated environment (#25)
Drop the usage of TVM to leverage the fact that we're running on isolated environments (GH hosted runners) and don't need the overhead of installing libraries for this purpose. In the foreseeable future, we might consider returning to using TVM if we adopt building images with a sharing resources approach, as with Jenkins. Still, with the current state of the implementation, the easiest maintainability route is to drop it altogether.